builders: Move experiment and ccu factory to builders module
